Fixed distro dection and bumped version to 1.0.2

+++ src/modules/10a-distro-ubuntu.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+test -e .distro && exit 1
+
 test -e /etc/issue || exit 1
 cat /etc/issue | grep -i "ubuntu" > /dev/null || exit 1
         
@@ -10,3 +12,4 @@ header      "Distribution Information"
 listitem    "Distribution"  "Ubuntu"
 listitem    "Release"       "$(cat /etc/issue)"
 
+touch .distro
